+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Format text routes (#447)
 - Add Lighthouse audits (#481)
 - Add Pa11y audits (#482)
+- Add Browserstack device compatibility testing (#480)
 - Comply with WCAG recommendations (#477)
 
 # 0.0.17 (staging release)